requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
cachetools==5.2.0
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
DateTime==4.4
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.0.4
django-heroku==0.3.1
djangorestframework==3.13.1
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.2
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
instagramy==4.5
kombu==5.2.4
lxml==4.9.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.1
playwright==1.22.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.1.0
pyparsing==3.0.9
pyppeteer-stealth==2.7.3
pytz==2022.1
redis==4.3.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tiktokscraper==0.1.0
tqdm==4.64.0
tweepy==4.10.0
typing_extensions==4.2.0
tzdata==2022.1
Unidecode==1.3.4
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
websockets==10.1
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.8.0
zope.interface==5.4.0
